ensuring chalk binary has read permissions
otherwise chalk cant read its own config and runs with defaults

also better sudo detection only when prefix is not owned by
current user

+++ setup.sh
@@ -68,6 +68,21 @@ function fatal {
     exit 1
 }
 
+function first_owner {
+    path=$1
+    while ! stat "$path" &> /dev/null; do
+        path=$(dirname "$path")
+    done
+    stat -c %u "$path"
+}
+
+function am_owner {
+    path=$1
+    uid=$(id -u)
+    path_uid=$(first_owner "$path")
+    [ "$uid" = "$path_uid" ]
+}
+
 function enable_debug {
     set -x
     log_level=trace
@@ -76,7 +91,7 @@ function enable_debug {
 
 # wrapper for calling chalk within the script
 function chalk {
-    $SUDO chmod +x "$chalk_path"
+    $SUDO chmod +xr "$chalk_path"
     timeout -s KILL 10s $SUDO "$chalk_path" --log-level=$log_level --skip-summary-report --skip-command-report "$@"
 }
 
@@ -111,7 +126,8 @@ function download_chalk {
     wget --quiet --directory-prefix=$TMP "$url"{,.sha256} || (
         fatal Could not download "$(chalk_version_name)". Are you sure this is a valid version?
     )
-    info Validating sha256 checksum
+    checksum=$(cat "$chalk_tmp.sha256")
+    info Validating sha256 checksum "${checksum%% *}"
     (
         cd $TMP
         $SHA256 -c "$chalk_tmp.sha256" || (
@@ -163,7 +179,7 @@ function add_line_to_config {
 
 # add lines to chalk config
 function add_lines_to_chalk {
-    config=$(mktemp -t chalk_XXXXXX.c4m)
+    config=$(mktemp -t chalk_XXXXXX).c4m
     chalk dump > "$config"
     for i; do
         add_line_to_config "$i" "$config"
@@ -199,12 +215,19 @@ function wrap_cmd {
     info Wrapping "$existing_path" command with chalk
 
     $SUDO mkdir -p "$(dirname "$chalkless_path")"
-    $SUDO ln "$existing_path" "$chalkless_path"
-    info Hardlinking "$chalkless_path" to "$existing_path"
+    if am_owner "$existing_path"; then
+        # hardlinking requires more permissions
+        # so only doing when owning file
+        info Hardlinking "$chalkless_path" to "$existing_path"
+        $SUDO ln "$existing_path" "$chalkless_path"
+    else
+        info Copying "$chalkless_path" to "$existing_path"
+        $SUDO cp "$existing_path" "$chalkless_path"
+    fi
 
     # create temporary chalk copy so that we can adjust its configuration
     # to be able to find the moved binary in the chalkless location
-    info Replacing "$chalked_path" with chalk
+    info Wrapping "$chalked_path" with chalk
     tmp=$(mktemp -t chalk.XXXXXX)
     $SUDO cp "$chalk_path" "$tmp"
     chalk_path=$tmp add_cmd_exe_to_config "$cmd" "$chalkless_path"
@@ -260,6 +283,14 @@ if [[ "$PATH" != ?(*:)"$prefix/bin"?(:*) ]]; then
 fi
 
 chalk_path=$prefix/bin/chalk
+
+if am_owner "$prefix"; then
+    SUDO=
+else
+    if [ -z "$SUDO" ]; then
+        fatal sudo is required to install chalk in "$prefix" as current user "'$(id -un)'" does not own it
+    fi
+fi
 
 if ! [ -f "$chalk_path" ] || [ -n "$overwrite" ]; then
     if [ -f "$chalk_path" ]; then
